backend/requirements.txt
# FastAPI backend
fastapi
uvicorn


# Environment variables
python-dotenv

# Authentication
passlib[bcrypt]
python-jose

# LangGraph + LLM
langgraph
langchain
langchain-google-genai

# LangGraph persistence (PostgreSQL)
langgraph-checkpoint-postgres
langsmith

sqlalchemy 
psycopg2-binary 
pydantic
email-validator
python-dateutil
python-jose[cryptography]